Improved the Configurations
* Allow "reloading" of the configurations along with observers to be
notified of the changes that have taken place since the last loading.
* Kotlin idiomatic way of accessing keys via property instead of func.
* Properties and Map based configurations will now convert numeric
strings to actual number types. Convienience capabilities.
* TqPropertiesConfig can now read file (or classpath if prepended with
classpath:).
* TqPropertiesConfig can now avoid throwing an exception on failure.
This is especially useful if chaining global and local settings where
one or the other might not exist when using the TqChainConfig.
* TqEnvConfig now loads everything up on init instead of on access.
This will now cause the keys to return the correct information.