Added insertReturnKey and transaction to return
* insertReturnKey() will do the same thing as insert but will also get
the generated key and return it as well in the form of a String. This
is because some systems will use a String as the id (such as Oracle).

* transaction will now allow returning a value back out. This prevents
the hackish paradigm of creating a var defaulting to null and then in
the transaction setting the desired value and so forth. Cleans up the
codebase a good bit.

* transaction will now default to commiting at the end instead of doing
a rollback. This helps keep things cleaner and easier to understand but
a commitOnCompletion flag can be passed in to revert it back to the
rollback state if desired.