Enhancements for use inside web container
Some tweaks were made so that when the Spark server is used within for
example a Tomcat container, that everything would work as expected.

The configurations will now check an alternative location for property
loading if it fails to be detected within the normal resource location.

Also opened up the NotAuthorizedException for downstream enterprise
application enhancement scenarios.

Also in order to clean up code usage when known immediate failure
situations happen, a validation enhancement to create an exception for
immediate throwing making the elvis operators usable for common case
scenarios.