Sparklin Enhancements
* Allow support for the 204 no content types by allowing Unit to be a
valid return type translating into no content being sent. This prevents
the use of ugly empty string return types I've seen some use.

* Support for non-json data to be utilized. For example image uploads
or xml data inputs which cannot be made into json now simply prevent
the use of the integrated validations but still allow for everything
else within the framework to work as desired.

* Helper for authz to work with variable arguments without having to
specify if it is all or nothing. Might still tweak or change this
around however as I'm not really happy with the way it feels right now.