Sparklin Validation Enhancements
* Changed default authz validation to be "all" instead of "or" default.
* Changed authz() to have the all/or flag after the varargs instead of
before. This makes it easier to consume the library from an end user
perspective but does requrire the use of the labelled parameter to
change the all/or flag.
* Fixed a bug in the "all" authz check when everything passed but still
caused you to be unauthorized.
* Added ifDefined and ifNotDefined validation helpers to allow for some
forking action if a parameter was or was not defined (which is not the
same thing as a null and there is no other way to determine this if a
proper data bean is used to back the request body).
* Added requiredOrNull validation to allow you to validate if something
is provided if it was provided at as value but also allow for null to
be an acceptable state. Basically reject if it is a blank string.
* Switched away from Spek back to Junit (Spek is awesome but not fully
integrated together with gradle, intellij, usability).