Cron Scheduling Support Added (TqCron)
This is a copy of the spring frameworks CronSequenceGenerator code but
tweaked a bit to be written in kotlin to support the kotlin idioms.

Added a NOTICE.txt to indicate that TqCron is not my own original work
and indicated where credit is due.

Also stripped out the Spek framework. While it is awesome and I really
want to be using it, there are just too many compromises I am having to
make in order to keep it going. Went back to junit 4 framework which is
just easier to do and has better tooling support for it as well.

Finally tweaked the TqRestClient in order to allow for customizations
with the object mappers if desired. This should address issues where
people may want to register custom transformation logic on objects or
reconfigure it but not have it be a global change.