jsonAs no longer nullable
* Also added email() validation check to ensure submitted data is a
valid email as much as possible.

The changing of jsonAs to return non-nullable values only is a change
made due to pragmatic reasons. In practice there is never really a use
for trying to transform nothing into objects and if it is, the code
does not really even attempt to handle the case as it is simple a bad
user request.

So simplify the lives of developers by not forcing the splattering of
the !! symbols all over the place causing numbness to their usage and
promoting bad best practices.