Tomcat Workaround Support
* Tomcat doesn't allow the content of a message to be read multiple
times. This causes various issues downstream when trying using sparklin
inside of a tomcat container. This update caches the body and provides
a way to obtain the originally read body. Unfortunately I could not
wrap the Request due to the protections that prevented me from working
around it via the normal wrapper techniques. So a new .rawBody() addon
extension is provided to gain access to the cached results and all the
features for validations and json transformations and such utilize the
cached body instead.
* Moved SparklinValidation out of core and into sparklin module. It was
accidentally moved to core when it should not have been.