Added TqSparklin as an efficiency for Sparklin
TqSparklin no longer always transforms payload data into objects and
does not place you within a Validation supported domain scope as well.
If a user wishes to reproduce the same capabilities as with Sparklin,
they may explicitly create a validation object to apply validations on
an as needed basis. This is more optimal now since validation is no
longer core to the Sparklin framework and can be used in a variety of
scenarios. This is for people who wish to use the framework but want
tighter control over what happens within it. So the built in authz
and validations are no longer there leaving it as core as possible
for anyone wishing to make customized handling of requests.

Also rolled back the spark version from 2.6.0 to 2.5.5 because of a bug
discovered with it under concurrency that caused the Exception Handling
to not be handled correctly.

Sparklin is still support and is the recommended implementation to use
because it does promote auth and validation as a 1st class citizen.

Will support both Sparklin and TqSpark for a while but I suspect that I
will want to try and unify the implementations even if it breaks the
backwards compatibility of one of them in order to try and get a single
solution in place that supports both use cases.

Squashed commit of the following:

commit 2cc0be5c6b11601e4211598f24825cf6c5598d46
Author: Sejal Patel <sejal@tekniq.io>
Date:   Sat Jul 8 14:58:22 2017 -0400

    Refactored from TqSpark to TqSparklin

    I just really liked the mix of Spark and Kotlin that made Sparklin. And
    the new TqSparklin is effectively still exactly the same as Sparklin
    except that it no longer forces object transformations for validation
    support. And because of the non-forced validation the integration of
    AuthorizationManager is no longer in place either. But both are still
    very trivial to utilize with explicit validation declaration instead of
    the forced and implicit validation declartions.

    Updated various README documentation as well.

commit 9495a0ee37f042b8bf466771f5a925dceddd4efc
Author: Sejal Patel <sejal@tekniq.io>
Date:   Sat Jul 8 13:13:01 2017 -0400

    Added TqSpark as an efficient version of Sparklin

    TqSpark no longer always transforms payload data into objects and does
    not place you within a Validation supported domain scope as well. If a
    user wishes to reproduce the same capabilities as with Sparklin, they
    may explicitly create a validation object (either a Map like before or
    using the same transformed data class they may normally work with) to
    apply validations on an as needed basis. This is more optimal now since
    validation is no longer core to the Sparklin framework and can be used
    in a variety of scenarios.

    Also rolled back the spark version from 2.6.0 to 2.5.5 because of a bug
    discovered with it under concurrency that caused the Exception Handling
    to not be handled correctly.

    With TqSpark the overhead of "validation" is no longer enforced on each
    call and the "confusion" of having magical validation is no longer in
    place either since it now must be explicitly defined.

    Will support both Sparklin and TqSpark for a while but I suspect that
    the Sparklin version may get deprecated at some point if TqSpark proves
    to be the more populate solution.