JDBC Transaction does explicit rollback on error
Seems that not all JDBC drivers properly rollback an uncommitted
connection when it is closed (a.k.a Oracle). So to help resolve issues
around a connection that has not been committed but a followup attempt
at something irrelevant complets also causing the previous uncommitted
transaction from completing, an explicit rollback is now being invoked
for any exception that takes place.