TqRestClient flag for allowing self signed certs
It's a common thing especially in enterprise code to communicate with
servers that are self-signed. You do still trust the server in question
but you don't wish to completely ignore all certificate checking across
all of the url connections. Thus this feature was added to allow for
improved usability of restful connections without exposing security
concerns across an application.

Also upgraded various dependencies to the latest.