Added ability to turn on logging
Also removed all the embedded test cases as they require the localhost
to have ssh server running on it and I also do not have a publicly
available always allowable system to ensure testing works. So if the
tests stayed as were, it wouldn't be possible for me to safely cut
releases even though the code works.

I validated everything once through and will need to eventually update
it so that it works like the rest of the test cases where it can be
disabled but executed at will to ensure that everything functions as
desired.