Validation Framework Enhancements
* Allow defaulting the validation src to null rather then explicitly
forcing the user to do so. This makes it a bit easier and prettier to
work with when tossing fixed validation errors.
* Fixed merging of validations to include the internal tested and passed
counters in order to provide accurate metrics.
* Added a new "with" support to allow for strongly typed checks to take
place using the also new "check" validation that allows freeform use of
anything returning a Boolean to be used as the determination of a
validation.
* Loosened up some validation internals to allow for subclassing of
validation and customizations to take place to better tune it for
specific project needs (like Javalin integration).

Also other changes include
* Upgraded to Kotlin 1.3.61 (from 1.3.50)
* Upgraded to Jackson 2.9.10 (from 2.9.8)
* Deprecated TqCache promoting use of the superior TqCaffeine instead.
Will probably merge tekniq-cache into the -core project at some point
and make the TqCache a wrapper around the TqCaffeine on the next feature
version bump.
* Added another classpath detection routine for the TqPropertiesConfig.
This is to help better support some of the changes that occur with
classpath support during Java 11+.