Enhanced JDBC extensions with CachedRowSet
Also added a "forEach" functionality on the ResultSet extension in order
to adapt the .next() style to a more consistent and kotlin friendly
approach. Plus it makes it a bit cleaner to use the CachedRowSet in a
practical setting.

This also allows use of a ResultSet even after the connection to the
database has been terminated. This is desired in many situations where
acting upon the data inline is not beneficial.