Improve TqConfig getInt/Long/Float/Double/Short
Previous it would not request the value from get() thus any child of the
config itself would not get evaluated for the value in order to be
converted. This left implementations that used a database as the backing
source for the data vulnerable to issues of the value not being found
correctly.

Also, once a value is found, it might be in a Double format which is not
an int causing a problem. So now all conversions will first go through a
transformation to a double format and from there to the desired final
output thus decimals in numbers like 3.0 will not cause the getInt to
break with a number format exception but instead return the correct
value of 3.

Another common check that people will want to do is ensuring that
content provided is a valid URL format. So added this capability in as
well.

Finally upgraded various dependencies to the latest compatible versions.