Improve JDBC Library Usability
When originally written .also and .let did not exist so there wasn't a
way to get a handle cleanly to the reference being acted upon. Thus
everything was centered around the .apply philosophy. Since then it has
become quite apparent that .apply is confusing and buggy especially with
nested complex subquery interactions.

This is a breaking compatibility modernization of the design approach to
enable future improvements.