Introduce Cryptography and Extensions
Upgraded multiple dependencies include the jackson dependency which
resolves a critical security threat.

In addition, API simplification of the TqConfig to eliminate the
parameterized definition of the default value used when missing as it is
almost always better to use the kotlin style for default value. Also,
pulling back null as a configuration value is extremely error prone and
unreliable depending on the backing source of that value so best to
eliminate it from contention as well.

Various cleanups and optimizations to the jdbc library. For example,
being able to do the insertReturnKey on the datasource level. Also
enabled anyone to use the apply params capability of the prepared
statement to better facilitate any compatible need with consistent rules
so timestamps and localdates and such are handled with minimal
frustration.

Improve documentation a bit to reflect some of the nice features that
have not been communicated in the past. Also to demonstrate some of the
new functionality that has been added.

Date math extensions added along with timezone safe time stripping.

Introduced a delegate for an initialization with default value. This is
to allow the setting of a value just once preventing its change. Thus
working much like the `val` declaration but without not with the hard
requirement of needing to be declared unless it is needed.