Streamline TqRest Performance
✨ Add support for streaming response data and optimizing for arrays
🔥 Remove RestLog support as it wasn't used in the wild much and
   hindered the ability to stream json array responses efficiently.

Overall not holding everything into memory first before transforming the
data improves the speed and lowers the memory requirements for existing
use cases by a tiny amount. But for cases where the response was an
array list, it creates a massive improvement in performance and reduces
the memory requirements when receiving large payloads.