Dependency Adjustments
Serious bug discovered in jackson-mapper 2.15.0 that causes it to not be
backwards compatible with code compiled using 2.14.x series which is
most of the world right now. Due to the non-functional backwards
compatibility, a downgrade of that dependency is needed for
compatibility with other libraries.

Upgrades to the latest bugfix/security releases of other dependencies
have been applied however.